more messing with BVH nodes
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@224 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ simpleraytracer/raymesh.cpp
@@ -681,7 +681,7 @@ void RayMesh::mergeVerticesWithSamePosAndNormal()
 
 			if(result == new_vert_indices.end())
 			{
-				new_vert_index = newverts.size();
+				new_vert_index = (unsigned int)newverts.size();
 				newverts.push_back(old_vert);
 				new_vert_indices.insert(std::make_pair(old_vert, new_vert_index));
 			}
